Fixed bug with follow_missile and balloon
The problem with the follow missile was solved by making the players
remove themselves from the group when they run out of energy.

The balloon issue was related to the fact that I was setting the rigid
body to a static body. If I set it to a kinematic body it still sends
the signal to the object to remove itself and give more energy to the
player. Perhaps the contact reporter doesn't work in static body mode?